core: only parse whitespace-related diffopts
A few months ago, I introduced a config to disable the a/ and b/ prefixes,
called diff.noprefix. Enabling this config caused problems with a variety of
core extensions such as mq and record, and when I dug in I found that they
really were doing the wrong thing -- that they should really be disabling all
format-changing diffopts even if they're enabled in ui. In order to do that I
introduced the difffeatureopts API, that lets extensions opt into features
they're interested in.

Make crecord use this API. The only options it is interested in are the
whitespace ones. nodates and git need to be explicitly enabled, so do that as
well.