Handle staged deletions correctly.
The diff Git produces has the file names in the header on both
sides even when the file is added or removed. To handle deletions
correctly, we need to check for a "deleted file mode" header.

Return None as a file name when then file is new or deleted,
and handle that further.