crecord: new keys g & G to navigate to the top and bottom respectively (#18)
This patch introduces two new keys 'g' and 'G' that helps to navigate to
the top and bottom of the file/hunk/line respectively. This is inline
with the shortcuts used in man, less, more and such tools that makes it
convenient to navigate swiftly. HOME and END keys also do the same.

'g' or HOME navigates to the top most file in the ncurses window.

'G' or END navigates to the bottom most file/hunk/line depending on the
whether the fold is active or not.

If the bottom most file is folded, it navigates to that file and stops
there. If the bottom most file is unfolded, it navigates to the bottom
most hunk in that file and stops there. If the bottom most hunk is
unfolded, it navigates to the bottom most line in that hunk.