Make sure staging actually works
The stage hotkey used to only set 'commit' option which was never
consulted, so the stage function was never called.

The stage function used to call git add with -N which only created
and entry in the index but didn’t stage the file itself.

When hitting 's' in commit mode and 'c' in staging mode, ask for
confirmation since it could be the user pressed the wrong key.

Fixes: #41